refactor: created package for domain
