+++ src/test/java/lotto/LottoPlayTest.java
@@ -1,9 +1,8 @@
 package lotto;
 
+import lotto.domain.LottoTickets;
 import org.junit.jupiter.api.Test;
 
-import java.util.ArrayList;
-import java.util.List;
 import java.util.Random;
 
 import static org.assertj.core.api.Assertions.assertThat;
